+++ lib/node.ts
@@ -546,9 +546,20 @@ export abstract class AstNode implements AstNodeBase {
 	 * Remove the current node
 	 *
 	 * 移除当前节点
+	 * @param ownLine whether to remove the current line if it is empty / 是否删除所在的空行
 	 */
-	remove(): void {
-		this.parentNode?.removeChild(this as AstNode as AstNodes);
+	remove(ownLine?: boolean): void {
+		const {parentNode, nextSibling, previousSibling} = this,
+			i = parentNode?.childNodes.indexOf(this as AstNode as AstNodes);
+		parentNode?.removeAt(i!);
+		if (
+			ownLine
+			&& parentNode?.getGaps(i! - 1) === 0
+			&& nextSibling?.type === 'text' && previousSibling?.type === 'text'
+			&& nextSibling.data.startsWith('\n') && previousSibling.data.endsWith('\n')
+		) {
+			nextSibling.deleteData(0, 1);
+		}
 	}
 
 	/**